+++ src/Instruction/Instructions/BVS.cpp
@@ -16,6 +16,6 @@ void BVS::run()
     if(!mIcpu->getFlag(Flag::V))
         return;
         
-    uint16_t result = mIcpu->getRegister(Register::PC) + addressingData.address;
+    uint16_t result = mIcpu->getRegister(Register::PC) + static_cast<int8_t>(addressingData.address);
     mIcpu->setRegister(Register::PC, result);
 }
